Make use of [loc_s] and [loc_e] in Coq fail messages to highlight the error.
Note: there might be a gap, in presence of utf8 chars, between the highlighted
zone and the zone Coq is referring to.
For example, "∅" is sent to Coq as "&#8709;" and while [len('∅') == 3],
[len('&#8709') == 7].

I am no python expert, so maybe I'm doing something wrong with my encode/decode,
but I've tried a few different configurations and the current one is the only
one which doesn't fail miserably.